refactor(headless-styles): create createClassProp helper and use in createIconButtonProps
