+++ packages/headless-styles/src/components/Button/types.ts
@@ -1,4 +1,5 @@
 import { ButtonHTMLAttributes, DetailedHTMLProps } from 'react'
+import type { Tech } from '../types'
 
 export interface ButtonOptions
   extends DetailedHTMLProps<
@@ -25,5 +26,4 @@ export type ButtonType = 'button' | 'submit' | 'reset' | undefined
 export type Kind = 'text' | 'textWeak' | 'weak' | 'medium' | 'strong'
 export type DangerKind = 'text' | 'medium' | 'strong'
 export type Size = 'xs' | 's' | 'm' | 'l'
-export type Tech = 'svelte' | ''
 export type Variant = 'default' | 'round'
